Move anchor code to it's own module and remove the setTimeOut code for code that does the scrolling after the page has finished loading.
