+++ resources/js/main.js
@@ -6,6 +6,7 @@ import './modules/formfilter';
 import './modules/slick';
 import './modules/magnific-popup';
 import './modules/share';
+import './modules/anchor';
 
 // Foundation Modules
 import './modules/accordion';
